requirements.txt
APScheduler==3.10.1
blinker==1.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
Flask-Mail==0.9.1
gunicorn==19.9.0
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.9.6
pymongo==3.10.1
python-http-client==3.2.7
pytz==2023.3
requests==2.30.0
shortuuid==1.0.1
six==1.16.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
Werkzeug==1.0.1
